requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
python-multipart>=0.0.6
pydantic>=2.5.3
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
alembic>=1.13.1
email-validator>=2.1.0
openai>=1.10.0
reportlab>=4.0.0
python-docx>=1.0.0
